% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-icd-parquet.R
\name{add_column_specific_diseases}
\alias{add_column_specific_diseases}
\title{Project specific diseases defined from ICD.}
\usage{
add_column_specific_diseases(data)
}
\arguments{
\item{data}{A `data.frame` type object (DuckDB, tibble, data.table). Assumes
DuckDB SQL connection.}
}
\value{
Same output as input.
}
\description{
Not all diseases we need for our projects are in DST's 99 overall classified
diseases.
}
\keyword{internal}
