% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{add_year_col}
\alias{add_year_col}
\title{Add year column to data if possible}
\usage{
add_year_col(data)
}
\arguments{
\item{data}{A data frame with a \code{source_file} column to extract year from.}
}
\value{
A data frame with a year column added if the year could be
determined.
}
\description{
Add year column to data if possible
}
\keyword{internal}
