% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-docs.R
\name{build_docs}
\alias{build_docs}
\title{Create the PDF files of the documentation.}
\usage{
build_docs()
}
\value{
Nothing, builds the PDF based on the Quarto book format.
}
\description{
Create the PDF files of the documentation.
}
\keyword{internal}
