% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_arrow_schema}
\alias{create_arrow_schema}
\title{Create a consistent Arrow schema from a data frame}
\usage{
create_arrow_schema(data)
}
\arguments{
\item{data}{A data frame to create a schema from.}
}
\value{
An Arrow schema with consistent types.
}
\description{
Maps R types to specific Arrow types to ensure consistent schemas across
chunks and files.
}
\keyword{internal}
