% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-icd-parquet.R
\name{create_lpr2_icd_parquet}
\alias{create_lpr2_icd_parquet}
\title{Convert ICD disease classification data from LPR2 to a partitioned Parquet database.}
\usage{
create_lpr2_icd_parquet(
  path_lpr_diag,
  path_lpr_adm,
  output_path,
  dst_icd_disease_groups
)
}
\arguments{
\item{path_lpr_diag}{The path to the LPR2 Diagnosis Parquet dataset.}

\item{path_lpr_adm}{The path to the LPR2 Admissions Parquet dataset.}

\item{output_path}{The new path to save the ICD data as a Parquet dataset.}

\item{dst_icd_disease_labels}{The disease labels given by DST for specific
ICD categories.}
}
\value{
A character vector.
}
\description{
Convert ICD disease classification data from LPR2 to a partitioned Parquet database.
}
\examples{
\dontrun{
lpr_adm_paths <- path_parquet_registers("708421") |>
  list_parquet_files() |>
  get_path_specific_database("lpr_adm")
lpr_diag_paths <- path_parquet_registers("708421") |>
  list_parquet_files() |>
  get_path_specific_database("lpr_diag")

create_lpr2_icd_parquet(
  path_lpr_adm = lpr_adm_paths,
  path_lpr_diag = lpr_diag_paths,
  path_alter_for_icd(lpr_diag_paths[1]),
  targets::tar_read(dst_icd_disease_labels)
)
}

}
\keyword{internal}
