% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-icd-parquet.R
\name{create_lpr3_icd_parquet}
\alias{create_lpr3_icd_parquet}
\title{Convert ICD disease classification data to a partitioned Parquet database.}
\usage{
create_lpr3_icd_parquet(
  path_diagnoser,
  path_kontakter,
  output_dir,
  dst_icd_disease_groups
)
}
\arguments{
\item{path_diagnoser}{The path to the LPR3 Diagnosis Parquet dataset.}

\item{path_kontakter}{The path to the LPR3 Admissions Parquet dataset.}

\item{output_dir}{The new path to save the ICD data as a Parquet dataset.}

\item{dst_icd_disease_labels}{The disease labels given by DST for specific
ICD categories.}
}
\value{
A character vector.
}
\description{
Convert ICD disease classification data to a partitioned Parquet database.
}
\examples{
\dontrun{
diagnoser_path <- path_parquet_external() |>
  list_parquet_files() |>
  get_path_specific_database("diagnoser")
kontakter_path <- path_parquet_external() |>
  list_parquet_files() |>
  get_path_specific_database("kontakter")

create_lpr3_icd_parquet(
  path_kontakter = kontakter_path,
  path_diagnoser = diagnoser_path,
  output_dir = as.character(fs::path(
    path_parquet_external("708421"),
    "icd"
  )),
  targets::tar_read(dst_icd_disease_labels)
)
}

}
\keyword{internal}
