% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_part_uuid}
\alias{create_part_uuid}
\title{Create UUID for partition part}
\usage{
create_part_uuid()
}
\value{
A character scalar with a UUID with a length of 6.
}
\description{
We're using shortened UUIDs instead of integers to avoid collisions when
converting registers in parallel.
}
\keyword{internal}
