% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastreg-package.R
\docType{package}
\name{fastreg-package}
\alias{fastreg}
\alias{fastreg-package}
\title{fastreg: Fast Conversion and Querying of Danish Registers with 'Parquet'}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Converts large Danish register files ('sas7bdat') into 'Parquet' format with year-based 'Hive' partitioning and chunked reading for larger-than-memory files. Supports parallel conversion with a 'targets' pipeline and reading those registers into 'DuckDB' tables for faster querying and analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dp-next.github.io/fastreg/} \url{https://github.com/dp-next/fastreg}
  \item Report bugs at \url{https://github.com/dp-next/fastreg/issues}
}

}
\author{
\strong{Maintainer}: Signe Kirk Brødbæk \email{signekb@clin.au.dk} (\href{https://orcid.org/0009-0000-2208-7088}{ORCID})

Authors:
\itemize{
  \item Signe Kirk Brødbæk \email{signekb@clin.au.dk} (\href{https://orcid.org/0009-0000-2208-7088}{ORCID})
  \item Luke Johnston \email{lwjohnst@gmail.com} (\href{https://orcid.org/0000-0003-4169-2616}{ORCID})
}

Other contributors:
\itemize{
  \item Steno Diabetes Center Aarhus [copyright holder]
  \item Aarhus University [copyright holder]
}

}
\keyword{internal}
