% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{get_conversion_log}
\alias{get_conversion_log}
\title{Get conversion logs from Targets pipeline output}
\usage{
get_conversion_log()
}
\value{
A tibble with the conversion logs.
}
\description{
Get conversion logs from Targets pipeline output
}
\examples{
\dontrun{
get_conversion_log()
}
}
