% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-helpers.R
\name{get_database_name}
\alias{get_database_name}
\title{Get the name of the database from the file name}
\usage{
get_database_name(path)
}
\arguments{
\item{path}{Path to the database file.}
}
\value{
A character with the database name.
}
\description{
Get the name of the database from the file name
}
\examples{
get_database_name("path/to/database.ext")
}
\keyword{internal}
