% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-helpers.R
\name{get_database_year}
\alias{get_database_year}
\title{Get the year of database from the file name}
\usage{
get_database_year(path)
}
\arguments{
\item{path}{Path to the database file.}
}
\value{
An integer with the year.
}
\description{
This function extracts the year of the database from the file name.
}
\examples{
get_database_year("path/to/database2025.sas")
}
\keyword{internal}
