% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-helpers.R
\name{get_database_year_external}
\alias{get_database_year_external}
\title{Get the years from the external database's name of the file path.}
\usage{
get_database_year_external(path)
}
\arguments{
\item{database_name}{The name of the database taken from the file path.}
}
\value{
A character vector of years.
}
\description{
Only for the external raw data files.
}
\keyword{internal}
