% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-helpers.R
\name{get_filename_no_ext}
\alias{get_filename_no_ext}
\title{Get the filename without its file extension}
\usage{
get_filename_no_ext(path)
}
\arguments{
\item{path}{Path to the file.}
}
\value{
A character with the filename without a file extension.
}
\description{
Get the filename without its file extension
}
\examples{
get_filename_no_ext("path/to/file name.ext")
}
