% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-icd-parquet.R
\name{get_icd_revision}
\alias{get_icd_revision}
\title{Determine whether ICD is 8 or 10 based on date.}
\usage{
get_icd_revision(date)
}
\arguments{
\item{date}{The date of diagnosis, when the code was used.}
}
\value{
A vector.
}
\description{
Determine whether ICD is 8 or 10 based on date.
}
\keyword{internal}
