% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{get_next_part}
\alias{get_next_part}
\title{Get the number of the next Parquet partition based on existing partitions.}
\usage{
get_next_part(path)
}
\arguments{
\item{path}{Path to (potentially) existing partitions.}
}
\value{
A padded string with the number of the next part. If no existing
parts are found at \code{path}, will return \code{0000}.
}
\description{
Get the number of the next Parquet partition based on existing partitions.
}
\keyword{internal}
