% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-helpers.R
\name{get_parquet_year}
\alias{get_parquet_year}
\title{Get the year of the parquet file from the file name}
\usage{
get_parquet_year(path)
}
\arguments{
\item{path}{Path to the parquet file.}
}
\value{
An integer with the year.
}
\description{
This function extracts the year of the database from the file name.
}
\examples{
get_parquet_year("path/to/year=2025/file.parquet")
}
