% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-helpers.R
\name{get_path_duplicates}
\alias{get_path_duplicates}
\title{Get paths with duplicate file names}
\usage{
get_path_duplicates(path)
}
\arguments{
\item{path}{Path to look for duplicate file names in.}
}
\value{
A character with duplicate paths, an empty character if no duplicates.
}
\description{
Get paths with duplicate file names
}
\examples{
get_path_duplicates(c("path/duplicate.parquet", "path/duplicate.parquet"))
get_path_duplicates(c("path/no/duplicate.parquet", "path/no/identical.parquet"))
}
\keyword{internal}
