% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-helpers.R
\name{get_path_no_duplicates}
\alias{get_path_no_duplicates}
\title{Get paths with no duplicate file names}
\usage{
get_path_no_duplicates(path)
}
\arguments{
\item{path}{Path to look for non-duplicate file names in.}
}
\value{
A character with unique paths, an empty character if duplicates.
}
\description{
Get paths with no duplicate file names
}
\examples{
get_path_no_duplicates(c("path/duplicate.parquet", "path/duplicate.parquet"))
get_path_no_duplicates(c("path/no/duplicate.parquet", "path/no/identical.parquet"))
}
