% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-helpers.R
\name{get_path_specific_database}
\alias{get_path_specific_database}
\title{Get path with specific database file}
\usage{
get_path_specific_database(path, name)
}
\arguments{
\item{path}{Path to get specific database from.}

\item{name}{Name of the specific database to look for.}
}
\value{
A character containing a path with the specific database.
}
\description{
Get path with specific database file
}
\examples{
get_path_specific_database("path/to/bef2025.sas", "bef")
get_path_specific_database("path/to/bef2025.sas", "lpr2")
}
\keyword{internal}
