% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-helpers.R
\name{get_path_with_year}
\alias{get_path_with_year}
\title{Get path with year in the file name}
\usage{
get_path_with_year(path)
}
\arguments{
\item{path}{Path to to get if there's a year in the file name.}
}
\value{
The path, if the file name contains a year.
}
\description{
Get path with year in the file name
}
\examples{
get_path_with_year(c("path/with/year2025", "path/without/year"))
}
