% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dir.R
\name{get_project_dir}
\alias{get_project_dir}
\title{Get a project directory by going up the directory tree}
\usage{
get_project_dir(project_id, path = ".")
}
\arguments{
\item{project_id}{The project id to get the directory of.}

\item{path}{The starting directory for the search. Defaults to the current
working directory.}
}
\value{
The absolute path to the matched directory.
}
\description{
Starting from \code{path}, this function walks up the directory tree to get the
first ancestor directory whose name matches \code{project_id}.
}
\examples{
temp_path <- fs::path_temp(123456, "subdir")
fs::dir_create(temp_path)
get_project_dir(project_id = 123456, path = temp_path)
}
