% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-helpers.R
\name{get_register_names}
\alias{get_register_names}
\title{Get the register names from file paths}
\usage{
get_register_names(path)
}
\arguments{
\item{path}{A character vector of one or more file paths.}
}
\value{
The file names from \code{path} with only letters (all non-letters
removed).
}
\description{
Removes all non-letters from the file names in \code{path}.
}
\keyword{internal}
