% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{get_year_from_col}
\alias{get_year_from_col}
\title{Get year from column values}
\usage{
get_year_from_col(data_col)
}
\arguments{
\item{data_col}{A character vector with file paths to extract year from.}
}
\value{
An integer vector with the extracted years, or \code{NA} if no year
found.
}
\description{
Get year from column values
}
\keyword{internal}
