% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{get_year_from_filename}
\alias{get_year_from_filename}
\title{Get year from file name}
\usage{
get_year_from_filename(file_path)
}
\arguments{
\item{file_path}{A character vector with file path to extract year from.}
}
\value{
An integer vector with the extracted years, or NA if no year
is found.
}
\description{
The year is determined as the first four consecutive numbers starting with
19 or 20 in the file name (i.e., years 1900-2099).
}
\keyword{internal}
