% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{is_same_register}
\alias{is_same_register}
\title{Check that all file paths are from the same register}
\usage{
is_same_register(file_paths)
}
\arguments{
\item{file_paths}{A character vector with paths to SAS register files.}
}
\value{
A logical that's TRUE if all \code{file_paths} point to files from the
same register, based on the file names.
}
\description{
Checks that all register names (file names without any non-letters) in
\code{file_paths} are identical, i.e., the registers have the same name.
}
\keyword{internal}
