% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{list_databases}
\alias{list_databases}
\title{Lists all the cleaned Parquet databases in the \code{cleaned-data} folder}
\usage{
list_databases(full_path = FALSE)
}
\arguments{
\item{full_path}{Full path to directory to list databases based on.}
}
\value{
A character vector.
}
\description{
Lists all the cleaned Parquet databases in the \code{cleaned-data} folder
}
\examples{
\dontrun{
list_databases()
}
}
