% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{list_dirs}
\alias{list_dirs}
\title{List directories at given path}
\usage{
list_dirs(path)
}
\arguments{
\item{path}{Path to list directories of.}
}
\value{
A character with directories at given path.
}
\description{
List directories at given path
}
\examples{
\dontrun{
list_dirs("raw")
}
}
