% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{list_misc_db}
\alias{list_misc_db}
\title{List all the available misc Parquet databases.}
\usage{
list_misc_db(full_path = FALSE)
}
\arguments{
\item{full_path}{Whether to list the full path to the database.}
}
\value{
Vector of file paths.
}
\description{
`r lifecycle::badge("deprecated")`
}
