% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_parquet_registers}
\alias{list_parquet_registers}
\title{List Parquet registers in a directory}
\usage{
list_parquet_registers(path)
}
\arguments{
\item{path}{The path to the directory to search for Parquet files.}
}
\value{
A character vector of paths to the Parquet files found.
}
\description{
This function lists all Parquet register files (with the extension \code{.parquet}
or \code{.parq} case-insensitively) in the specified directory and its
subdirectories.
}
\examples{
list_parquet_registers(fs::path_package("registers2parquet", "extdata"))
}
