% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{list_registers}
\alias{list_registers}
\title{List all the available registers.}
\usage{
list_registers(full_path = FALSE)
}
\arguments{
\item{full_path}{Whether to list the full path to the register.}
}
\value{
Vector of file paths.
}
\description{
`r lifecycle::badge("deprecated")`
}
