% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-parquet.R
\name{load_database}
\alias{load_database}
\title{Load a specific Parquet database into the R session}
\usage{
load_database(name)
}
\arguments{
\item{name}{Name of the database to load.}
}
\value{
The database as an Arrow dataset.
}
\description{
The loaded database will be converted into DuckDB so it can be used via
\code{\link[arrow:to_duckdb]{arrow::to_duckdb()}}.
}
