% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-parquet.R
\name{load_misc_db}
\alias{load_misc_db}
\title{Load a specific misc database into the R session}
\usage{
load_misc_db(name)
}
\arguments{
\item{name}{Name of the register to load.}
}
\value{
The register dataset as an Arrow dataset.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
The loaded database will be converted into DuckDB so it can be used via
\code{\link[arrow:to_duckdb]{arrow::to_duckdb()}}.
}
