% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_schema}
\alias{log_schema}
\title{Log schema information}
\usage{
log_schema(register_log)
}
\arguments{
\item{register_log}{A tibble returned by \code{\link[=convert]{convert()}}, filtered to only
contain rows from a single register.}
}
\value{
A \code{s3} object \code{fastreg_schema} with the structure:
\itemize{
\item description: Character describing the schemas (diff or no diff).
\item reference_schema: Tibble with the most common schema.
\item diff_tables: A list with tibble(s) showing the columns that differ from
the reference schema. If there's many files with schema differences,
there will be multiple tables, for printing purposes.
}
}
\description{
Turns the log schema information into a section that compares the schemas
within one register. Finds the most common schema and if there's differences
between schemas, it reports these differences.
}
\examples{
sas_file <- fs::path_package("fastreg", "extdata", "test.sas7bdat")
conversion_log <- convert(sas_file, output_dir = fs::path_temp("output"))
log_schema(conversion_log)
}
