% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{path_alter_filename_as_dir}
\alias{path_alter_filename_as_dir}
\title{Convert path to end with `filename/`}
\usage{
path_alter_filename_as_dir(path)
}
\arguments{
\item{path}{A file path.}
}
\value{
A character vector.
}
\description{
Convert path to end with `filename/`
}
\examples{
fs::file_temp(pattern = "database", ext = ".sas7bdat") |>
  path_alter_filename_as_dir()
}
