% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{path_alter_filename_year_as_dir}
\alias{path_alter_filename_year_as_dir}
\title{Convert file name of a path to end in `/year=YYYY`}
\usage{
path_alter_filename_year_as_dir(path)
}
\arguments{
\item{path}{A file path.}
}
\value{
A character vector.
}
\description{
To follow the Parquet partitioning style (`{input_path_dir}/year=####/`).
}
\examples{
fs::file_temp(pattern = "database2020-", ext = ".sas7bdat") |>
  path_alter_filename_year_as_dir()
}
