% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{path_alter_to_cleaned_dir}
\alias{path_alter_to_cleaned_dir}
\title{Convert path to cleaned directory.}
\usage{
path_alter_to_cleaned_dir(path, dir)
}
\arguments{
\item{path}{Path to convert.}

\item{dir}{Directory to output to.}
}
\value{
The clean directory.
}
\description{
A cleaned directory in this context means a Parquet partition.
}
