% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{path_alter_to_output_parquet_partition}
\alias{path_alter_to_output_parquet_partition}
\title{Convert the path to represent a Parquet Partition in another directory}
\usage{
path_alter_to_output_parquet_partition(path, output_dir)
}
\arguments{
\item{path}{A file path.}

\item{output_dir}{New directory the file should be in.}
}
\value{
A character vector.
}
\description{
Converts the file path to the pattern:
`{output_dir}/{file_name}/part-0.parquet`, since this is the style used to
tell Parquet the file is part of a partition. This includes converting to a
year partition as `{output_dir}/{file_name}/style={year}/part-0.parquet`.
}
\examples{
fs::path_temp("database.sas7bdat") |>
  path_alter_to_output_parquet_partition(fs::path_temp("new-dir"))
fs::file_temp(pattern = "database2020-", ext = ".sas7bdat") |>
  path_alter_to_output_parquet_partition(fs::path_temp("new-dir"))
}
