% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{path_as_df}
\alias{path_as_df}
\title{Create dataframe with path and file name}
\usage{
path_as_df(path)
}
\arguments{
\item{path}{Path to data.}
}
\value{
A tibble with the path(s).
}
\description{
Create a dataframe (tibble) with columns "path" and "file", with the
full path and file name respectively.
}
\examples{
path_as_df(c("path/data.parquet", "path/another.parquet"))
}
