% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{path_duplicates_as_list}
\alias{path_duplicates_as_list}
\title{Get duplicate paths as a list}
\usage{
path_duplicates_as_list(path)
}
\arguments{
\item{path}{Paths to list.}
}
\value{
A list with input path(s) where duplicate file names are grouped in the same list.
}
\description{
Get duplicate paths as a list
}
\examples{
path_duplicates_as_list(c("path/data.parquet", "path/another.parquet", "path/to/another.parquet"))
}
