% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_eksterne_dir}
\alias{path_eksterne_dir}
\title{Path to external directory}
\usage{
path_eksterne_dir(id = "708421")
}
\arguments{
\item{id}{The project ID.}
}
\value{
Path to the external directory.
}
\description{
Path to external directory
}
\examples{
\dontrun{
path_eksterne_dir()
}
}
