% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{path_ext_set_parquet_partition}
\alias{path_ext_set_parquet_partition}
\title{Convert file path to Parquet Partition}
\usage{
path_ext_set_parquet_partition(path)
}
\arguments{
\item{path}{A file path.}
}
\value{
A character vector.
}
\description{
Convert file path to Parquet Partition
}
\examples{
fs::file_temp(ext = ".sas7bdat") |> path_ext_set_parquet_partition()
}
