% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_grunddata_dir}
\alias{path_grunddata_dir}
\title{Path to "grunddata" directory}
\usage{
path_grunddata_dir(id = "708421")
}
\arguments{
\item{id}{The project ID.}
}
\value{
Path to the "grunddata" directory within the project.
}
\description{
Path to "grunddata" directory
}
\examples{
\dontrun{
path_grunddata_dir()
}
}
