% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_parquet_dirs}
\alias{path_parquet_dirs}
\title{Path to Parquet directory}
\usage{
path_parquet_dirs(id = "708421")
}
\arguments{
\item{id}{The project ID.}
}
\value{
Path to the Parquet directory.
}
\description{
Path to Parquet directory
}
\examples{
\dontrun{
path_parquet_dirs()
}
}
