% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_parquet_external}
\alias{path_parquet_external}
\title{Path to external Parquet files}
\usage{
path_parquet_external(id = "708421")
}
\arguments{
\item{id}{The project ID.}
}
\value{
Path to external Parquet files within the project.
}
\description{
Path to external Parquet files
}
\examples{
\dontrun{
path_parquet_external()
}
}
