% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_parquet_registers}
\alias{path_parquet_registers}
\title{Path to Parquet registers}
\usage{
path_parquet_registers(id = "708421")
}
\arguments{
\item{id}{The project ID.}
}
\value{
Path to the Parquet registers within the project.
}
\description{
Path to Parquet registers
}
\examples{
\dontrun{
path_parquet_registers()
}
}
