% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_population_file}
\alias{path_population_file}
\title{Path to population file}
\usage{
path_population_file(id = "708421")
}
\arguments{
\item{id}{The project ID.}
}
\value{
The path to the population file within the project.
}
\description{
Path to population file
}
\examples{
\dontrun{
path_population_file()
}
}
