% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_rawdata}
\alias{path_rawdata}
\title{Path to rawdata directory}
\usage{
path_rawdata(id)
}
\arguments{
\item{id}{The project ID.}
}
\value{
Path to the rawdata directory within the project.
}
\description{
Path to rawdata directory
}
\examples{
path_rawdata(1)
}
