% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{path_set_dir}
\alias{path_set_dir}
\title{Alter the path of a file to a Parquet partition in another directory.}
\usage{
path_set_dir(path, output_dir)
}
\arguments{
\item{path}{A file path.}

\item{output_dir}{New directory the file should be in.}
}
\value{
A character vector.
}
\description{
Alter the path of a file to a Parquet partition in another directory.
}
\examples{
fs::file_temp(ext = ".sas7bdat") |> path_set_dir(fs::path_temp())
}
\keyword{internal}
