% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_subdir}
\alias{path_subdir}
\title{Path to subdirectory}
\usage{
path_subdir(path, dir)
}
\arguments{
\item{path}{Path to return subdirectory of.}

\item{dir}{Subdirectory to add to path.}
}
\value{
Path with given subdirectory of given path.
}
\description{
Path to subdirectory
}
\examples{
\dontrun{
path_subdir("rawdata", "cleaned")
}
}
