% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_workdata}
\alias{path_workdata}
\title{Path to workdata directory}
\usage{
path_workdata(id)
}
\arguments{
\item{id}{The project ID.}
}
\value{
Path to the workdata directory within the project.
}
\description{
Path to workdata directory
}
\examples{
path_workdata(1)
}
