% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\title{List important paths for projects on DST.}
\arguments{
\item{id}{The Project ID.}

\item{dir}{The directory name to get the path for.}
}
\description{
List important paths for projects on DST.
}
\examples{
path_rawdata("708421")
path_population_file("708421")
path_grunddata_dir()
path_eksterne_dir()
path_parquet_dirs()
path_sas_formats()
}
