% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{print.fastreg_schema}
\alias{print.fastreg_schema}
\title{Print method for the S3 class fastreg_schema}
\usage{
\method{print}{fastreg_schema}(schema, ...)
}
\arguments{
\item{schema}{The fastreg_schema returned by \code{\link[=log_schema]{log_schema()}}.}

\item{...}{Not used; included for S3 method compatibility.}
}
\value{
The schema invisibly.
}
\description{
Print method for the S3 class fastreg_schema
}
\examples{
sas_file <- fs::path_package("fastreg", "extdata", "test.sas7bdat")
conversion_log <- convert(sas_file, output_dir = fs::path_temp("output"))
log_schema(conversion_log) |> print()
}
