% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{print_log_schema}
\alias{print_log_schema}
\title{Print log schema comparison}
\usage{
print_log_schema(register_log)
}
\arguments{
\item{register_log}{A tibble returned by \code{\link[=convert]{convert()}}, filtered to only
contain rows from a single register.}
}
\value{
register_log invisibly.
}
\description{
Prints the log schema information in a section that compares the schemas
within one register. Finds the most common schema and if there's differences
between schemas, it prints these differences.
}
\examples{
sas_file <- fs::path_package("fastreg", "extdata", "test.sas7bdat")
log <- convert(sas_file, output_dir = fs::path_temp("output"))
print_log_schema(log)
}
