% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-sample.R
\name{randomly_sample}
\alias{randomly_sample}
\title{Randomly sample PNR values to create a smaller database to use to test code on.}
\usage{
randomly_sample(population_db, register_db, n = 1e+06)
}
\arguments{
\item{population_db}{The population database that defines the people found in our DST project.}

\item{register_db}{The register you want to randomly sample from to make smaller.}

\item{n}{Sample size to use. Default is a million (1e6).}
}
\value{
Either a tibble or duckdb connection, depending on input databases.
}
\description{
Inside the function we keep only rows where `cprtjek` is 0, since that means
that there isn't a problem with the person's data or number (as defined by DST).
We also drop missing PNR numbers (there aren't many, just to be sure).
}
