% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-icd-parquet.R
\name{read_diagnoser}
\alias{read_diagnoser}
\title{Load in the LPR3 diagnose codes}
\usage{
read_diagnoser(path)
}
\arguments{
\item{path}{Path to file(s) for the LPR3 diagnoser kodes}
}
\value{
A DuckDB object.
}
\description{
Load in the LPR3 diagnose codes
}
\examples{
\dontrun{
path_parquet_external() |>
  list_parquet_files() |>
  get_path_specific_database("diagnoser") |>
  read_diagnoser()
}
}
