% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-icd-labels.R
\name{read_icd_disease_labels}
\alias{read_icd_disease_labels}
\title{Read both ICD8 and ICD10 labeling data as a tibble.}
\usage{
read_icd_disease_labels()
}
\value{
A tibble.
}
\description{
Calls [read_specific_icd_labels()] for both ICD8 and ICD10, then joins both.
}
\examples{

read_icd_disease_labels()

}
