% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-icd-parquet.R
\name{read_lpr2}
\alias{read_lpr2}
\title{Reads in the LPR2 registers and combines them.}
\usage{
read_lpr2(path_lpr_diag, path_lpr_adm)
}
\arguments{
\item{path_lpr_diag}{Path to the LPR2 diagnosis data files.}

\item{path_lpr_adm}{Path to the LPR2 diagnosis data files.}
}
\value{
A DuckDB object.
}
\description{
Reads in the LPR2 registers and combines them.
}
\examples{
\dontrun{
read_lpr2(
  path_parquet_registers() |>
    list_parquet_files() |>
    get_path_specific_database("lpr_diag"),
  path_parquet_registers() |>
    list_parquet_files() |>
    get_path_specific_database("lpr_adm")
)
}
}
\keyword{internal}
