% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-icd-parquet.R
\name{read_lpr3}
\alias{read_lpr3}
\title{Read in the LPR3 registers.}
\usage{
read_lpr3(path_diagnoser, path_kontakter)
}
\arguments{
\item{path_diagnoser}{Path to the diagnoser data file.}

\item{path_kontakter}{Path to the kontakter data file.}
}
\value{
A DuckDB object.
}
\description{
Read in the LPR3 registers.
}
\examples{
\dontrun{
read_lpr3(
  path_parquet_external() |>
    list_parquet_files() |>
    get_path_specific_database("diagnoser"),
  path_parquet_external() |>
    list_parquet_files() |>
    get_path_specific_database("kontakter")
)
}
}
