% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-icd-parquet.R
\name{read_lpr_adm}
\alias{read_lpr_adm}
\title{Read the LPR Admissions Parquet database.}
\usage{
read_lpr_adm(path)
}
\arguments{
\item{path}{A vector of Parquet paths for the `lpr_adm`.}
}
\value{
A DuckDB database connection.
}
\description{
Read the LPR Admissions Parquet database.
}
\examples{
\dontrun{
path_parquet_registers() |>
  list_parquet_files() |>
  get_path_specific_database("lpr_adm") |>
  read_lpr_adm()
}

}
\keyword{internal}
