% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-icd-parquet.R
\name{read_lpr_diag}
\alias{read_lpr_diag}
\title{Read the LPR Diagnosis Parquet database.}
\usage{
read_lpr_diag(path)
}
\arguments{
\item{path}{A vector of Parquet paths for the `lpr_diag`.}
}
\value{
A DuckDB database connection.
}
\description{
Read the LPR Diagnosis Parquet database.
}
\examples{
\dontrun{
path_parquet_registers() |>
  list_parquet_files() |>
  get_path_specific_database("lpr_diag") |>
  read_lpr_diag()
}

}
\keyword{internal}
