% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_parquet}
\alias{read_parquet}
\alias{read_parquet_partition}
\alias{read_parquet_file}
\title{Read a single Parquet file or a partitioned dataset as DuckDB table}
\usage{
read_parquet_partition(path)

read_parquet_file(path)
}
\arguments{
\item{path}{Path to a directory with the Parquet files within or a path to a
Parquet file.}
}
\value{
A DuckDB table.
}
\description{
This is useful when the \code{\link[=read_register]{read_register()}} incorrectly guesses or can't find
the register.
}
\section{Functions}{
\itemize{
\item \code{read_parquet_partition()}: Reads a Parquet partitioned directory.

\item \code{read_parquet_file()}: Reads a single Parquet file.

}}
