% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_parquet_file_as_duckdb}
\alias{read_parquet_file_as_duckdb}
\title{Read a Parquet file as DuckDB table}
\usage{
read_parquet_file_as_duckdb(file_path)
}
\arguments{
\item{file_path}{A character scalar with the path to the Parquet file.}
}
\value{
The register as a DuckDB table.
}
\description{
Read a Parquet file as DuckDB table
}
\keyword{internal}
