% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-parquet.R
\name{read_parquet_partition}
\alias{read_parquet_partition}
\title{Read an individual Parquet partitioned database.}
\usage{
read_parquet_partition(dir)
}
\arguments{
\item{dir}{The folder that holds the Parquet database.}
}
\value{
A DuckDB database connection.
}
\description{
Read an individual Parquet partitioned database.
}
\keyword{internal}
