% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_parquet_partition_as_duckdb}
\alias{read_parquet_partition_as_duckdb}
\title{Read a partitioned Parquet register as DuckDB table}
\usage{
read_parquet_partition_as_duckdb(dir_path)
}
\arguments{
\item{dir_path}{A character scalar with the path to the Parquet register
directory.}
}
\value{
The register as a DuckDB table.
}
\description{
Read a partitioned Parquet register as DuckDB table
}
\keyword{internal}
