% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{read_sas_files}
\alias{read_sas_files}
\title{Read SAS files}
\usage{
read_sas_files(path)
}
\arguments{
\item{path}{A character vector with the absolute path to the SAS file(s).}
}
\value{
A data frame with the contents of the SAS file(s) plus a
\code{source_file} column.
}
\description{
This function reads one or more SAS files and adds a \code{source_file} column
indicating the file each row came from.
}
\keyword{internal}
