% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-icd-labels.R
\name{read_specific_icd_disease_labels}
\alias{read_specific_icd_disease_labels}
\title{Reads one individual ICD label file from DST's documentation.}
\usage{
read_specific_icd_disease_labels(icd_revision = c("ICD8", "ICD10"))
}
\arguments{
\item{icd_revision}{Either ICD8 or ICD10.}
}
\value{
A tibble.
}
\description{
This imports *either* the ICD8 or ICD10 "L1L2_KT" files in DST's variable
format folder. L1L2 stands for DST's "99 disease groupings". There's another
one for "23 disease groupings", but we'll use this one to have a bit more
resolution at the disease level.
}
\examples{

read_specific_icd_disease_labels("ICD8")
read_specific_icd_disease_labels("ICD10")

}
\keyword{internal}
