% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas-to-parquet.R
\name{sas_to_parquet}
\alias{sas_to_parquet}
\title{Read the raw SAS database file and save to a Parquet format.}
\usage{
sas_to_parquet(input_path, output_path)
}
\arguments{
\item{input_path}{The path to the raw SAS file.}

\item{output_path}{The path to save the Parquet file.}
}
\value{
Returns a character vector of the created Parquet files (from
  `output_path`), so as to be designed to work with targets.
}
\description{
Parquet is a format for saving data in an efficient way and that allows you
to easily import it either with `arrow::read_arrow()` function or with
any SQL-based language. In this case, I'm recommending using DuckDB
via `arrow::to_duckdb()` since it is a SQL language that is fast and is
designed for data analysis tasks.
}
