% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_diagnoser_register}
\alias{simulate_diagnoser_register}
\title{Create example diagnoser registers}
\usage{
simulate_diagnoser_register(n = 1000)
}
\arguments{
\item{n}{Number of rows to simulate per simulated diagnoser tibble.}
}
\value{
A named list with two tibbles containing simulated diagnoser
registers.
}
\description{
The data is simulated using \code{osdc::simulate_registers()}. It's used
in vignettes and tests.
}
