% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_kontakter_register}
\alias{simulate_kontakter_register}
\title{Create example kontakter registers}
\usage{
simulate_kontakter_register(n = 1000)
}
\arguments{
\item{n}{Number of rows to simulate per simulated kontakter tibble.}
}
\value{
A named list with four tibbles containing simulated kontakter
registers.
}
\description{
The data is simulated using \code{osdc::simulate_registers()}. It's used
in vignettes and tests.
}
