% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_file_in_chunks}
\alias{convert_file_in_chunks}
\title{Convert a single register SAS file to Parquet in chunks}
\usage{
convert_file_in_chunks(file_path, output_dir, chunk_size = 10000000L)
}
\arguments{
\item{file_path}{A character scalar with the absolute path to a single SAS file.}

\item{output_dir}{A character scalar with the path to the directory to save
the output Parquet file to. Should include the register name as the last
part of the path. E.g., \verb{path/to/register_name/}.}

\item{chunk_size}{An integer scalar indicating the number of rows to read
at a time from the SAS files. Defaults to 10,000,000.}
}
\value{
Path to the partition.
}
\description{
Convert a single register SAS file to Parquet in chunks
}
\keyword{internal}
