% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_to_parquet}
\alias{convert_to_parquet}
\title{Convert register SAS file(s) and save to Parquet format}
\usage{
convert_to_parquet(path, output_path)
}
\arguments{
\item{path}{A character vector with the absolute path to the register SAS
file(s).}

\item{output_path}{A character scalar with the path to the directory to save
the output Parquet file to. Should include the register name as the last
part of the path. E.g., \verb{path/to/register_name/}.}
}
\value{
Returns a character scalar with the path to the created Parquet
file(s) (\code{output_path}), so it can be used in a \href{https://books.ropensci.org/targets/}{targets} pipeline.
}
\description{
If multiple paths are given, the function looks for a year (4 digits) in the
file names to use the year as partition, see \code{vignettes("design")} for more
information about the partitioning. If a year is found, the data is saved
partitioned by year in the output directory, e.g.,
\verb{path/to/register_name/year=2020/part-0.parquet}.

If no year can be found or only one path is given, the data is converted
without partitioning and saved as a Parquet file with the name specified in
the output path. E.g., if output_path is \code{path/to/register}, the Parquet file
will be saved as \code{path/to/register.parquet}.

If any duplicate rows are found, they are deduplicated before saving to
Parquet. If duplicate rows are found in multiple source files, the row from
the file that appears first in \code{path} is kept. Rows that are almost identical
across different files (e.g. different years) but that have a difference in
values are kept, as determining which is the correct value requires domain
knowledge.
}
\examples{
\dontrun{
convert_to_parquet(
  list_sas_registers(project_id="202020")),
  "output/path/to/register_name")
}
}
