% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_project_id}
\alias{get_project_id}
\title{Get (or guess) the project ID from the current working directory}
\usage{
get_project_id()
}
\value{
The project ID as a character string.
}
\description{
Get (or guess) the project ID from the current working directory
}
