% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-helpers.R
\name{get_register_name}
\alias{get_register_name}
\title{Get register name from a group of file paths}
\usage{
get_register_name(path)
}
\arguments{
\item{path}{A character vector of one or more paths from the same register.}
}
\value{
A character scalar with the register name.
}
\description{
Extracts the register name from the path in a group. Intended for use
with groups created by \code{\link[=split_paths_by_register]{split_paths_by_register()}} in the targets template.
}
\examples{
path <- c("data/bef2020.sas7bdat", "data/bef2021.sas7bdat")
get_register_name(path)
}
