% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-helpers.R
\name{get_register_names}
\alias{get_register_names}
\title{Get the register names from file paths}
\usage{
get_register_names(paths)
}
\arguments{
\item{paths}{Character vector with file paths.}
}
\value{
The file names from \code{paths} with only letters (all non-letter
removed).
}
\description{
Removes all non-letters from the file names in \code{paths}.
}
\keyword{internal}
