% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{is_same_register}
\alias{is_same_register}
\title{Check that all paths are from the same register.}
\usage{
is_same_register(paths)
}
\arguments{
\item{paths}{A character vector with paths to SAS registers.}
}
\value{
A logical that's TRUE if all paths point to files from the same
register, based on the file names.
}
\description{
Removes all non-letters from the file names in paths and checks that the
remaining characters are identical, i.e., the registers have the same name.
}
\keyword{internal}
