% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{list_databases}
\alias{list_databases}
\title{Lists all the cleaned Parquet databases in the `cleaned-data` folder}
\usage{
list_databases(full_path = FALSE)
}
\value{
A character vector.
}
\description{
Lists all the cleaned Parquet databases in the `cleaned-data` folder
}
\examples{
\dontrun{
list_databases()
}
}
