% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_parquet}
\alias{list_parquet}
\alias{list_parquet_datasets}
\alias{list_parquet_files}
\title{List Parquet datasets or files in a project}
\usage{
list_parquet_datasets()

list_parquet_files()
}
\value{
The path(s) to the Parquet datasets (as directories) or files.
}
\description{
Only lists Parquet files that end in \verb{part-*.parquet}. For datasets,
it will only look for Parquet files with a \code{year=YYYY} in its path.
This function will search the whole system for the project ID, so it might
be slow sometimes.
}
\section{Functions}{
\itemize{
\item \code{list_parquet_datasets()}: List all Parquet (Hive partitioned by year) datasets.

\item \code{list_parquet_files()}: List all Parquet files within a project.

}}
