% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_parquet_files}
\alias{list_parquet_files}
\title{List Parquet files in a directory}
\usage{
list_parquet_files(path)
}
\arguments{
\item{path}{Directory to search.}
}
\value{
The path(s) to the found SAS files.
}
\description{
Lists all Parquet register files (with the extension \code{.parquet}
or \code{.parq} case-insensitively) in the specified directory and its
subdirectories.
}
\examples{
list_parquet_files(fs::path_package("fastreg", "extdata"))
}
