% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{list_parquet_files}
\alias{list_parquet_files}
\title{List all the Parquet database files recursively in a folder.}
\usage{
list_parquet_files(dir)
}
\arguments{
\item{dir}{Directory to list (recursively) all Parquet files.}
}
\value{
A character vector.
}
\description{
List all the Parquet database files recursively in a folder.
}
\examples{

list_parquet_files(path_parquet_dirs()[1])
list_parquet_files(path_parquet_dirs()[2])

}
