% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_sas_files}
\alias{list_sas_files}
\title{List SAS files in a directory}
\usage{
list_sas_files(path)
}
\arguments{
\item{path}{A character scalar with the path to the directory to search.}
}
\value{
A character vector of paths to the SAS files found.
}
\description{
Lists all SAS register files (with the extension \code{.sas7bdat}
case-insensitively) in the specified directory and its subdirectories.
}
\examples{
list_sas_files(fs::path_package("fastreg", "extdata"))
}
