% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{list_sas_files}
\alias{list_sas_files}
\title{List all the SAS register databases recursively in a folder}
\usage{
list_sas_files(dir)
}
\arguments{
\item{dir}{Directory to list (recursively) all SAS database files (e.g. in
`Grunddata/`).}
}
\value{
A character vector of file paths.
}
\description{
List all the SAS register databases recursively in a folder
}
\examples{
\dontrun{
list_sas_files(path_grunddata_dir())
list_sas_files(path_eksterne_dir())
}
}
