% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-parquet.R
\name{load_database}
\alias{load_database}
\title{Load a specific Parquet database into the R session.}
\usage{
load_database(name)
}
\arguments{
\item{name}{Name of the database to load.}
}
\value{
Opens the database as a Arrow dataset.
}
\description{
The loaded database will be converted into DuckDB so it can use it, via
[arrow::to_duckdb()].
}
