% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-parquet.R
\name{load_register}
\alias{load_register}
\title{Load a specific register into the R session}
\usage{
load_register(name)
}
\arguments{
\item{name}{Name of the register to load.}
}
\value{
The register dataset as an Arrow dataset.
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
The loaded register will be converted into DuckDB so it can be used via
[arrow::to_duckdb()].
}
