% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_as_table}
\alias{log_as_table}
\title{Log chunk information as a table}
\usage{
log_as_table(register_log)
}
\arguments{
\item{register_log}{A tibble returned by \code{\link[=convert]{convert()}}, filtered to only
contain rows from a single register.}
}
\value{
A \code{knitr_kable} table.
}
\description{
Turns the log information returned by \code{\link[=convert]{convert()}} into a pretty
table, showing relative input/output paths and row counts.
}
\examples{
sas_file <- fs::path_package("fastreg", "extdata", "test.sas7bdat")
conversion_log <- convert(sas_file, output_dir = fs::path_temp("output"))
log_as_table(conversion_log)
}
