% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_register}
\alias{read_register}
\title{Read a Parquet register}
\usage{
read_register(name)
}
\arguments{
\item{name}{Name of the register. See a list of available registers with
\code{\link[=list_parquet_datasets]{list_parquet_datasets()}}.}
}
\value{
A DuckDB table.
}
\description{
This function uses the options \code{fastreg.project_rawdata_dir} and
\code{fastreg.project_workdata_dir} when set in \code{\link[=options]{options()}} or will try to guess
the path by using the project ID and the base directories
\verb{E:/<project-id>/rawdata/} and \verb{E:/<project-id>/workdata/}. Will only read
Parquet datasets (those that are partitioned with the pattern \verb{year=}). If
this function doesn't work, use \code{\link[=read_parquet_partition]{read_parquet_partition()}} or
\code{\link[=read_parquet_file]{read_parquet_file()}} instead.
}
