% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_register}
\alias{read_register}
\title{Read a Parquet register}
\usage{
read_register(path)
}
\arguments{
\item{path}{A character scalar with the path to the Parquet register.}
}
\value{
The register as a DuckDB table.
}
\description{
If you want to read a partitioned Parquet register, provide the path to the
directory (e.g., \verb{path/to/parquet/register/}).
If you want to read a single Parquet file, provide the path to the file
(e.g., \code{path/to/parquet/register.parquet}).
}
\examples{
read_register(fs::path_package(
  "registers2parquet",
  "extdata",
  "test_register.parquet"
))
}
