% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{read_sas_files}
\alias{read_sas_files}
\title{Read SAS files}
\usage{
read_sas_files(path)
}
\arguments{
\item{path}{A character vector with the absolute path to the SAS file(s).}
}
\value{
A data frame with the contents of the SAS file(s) plus a
\code{source_file} column.
}
\description{
This function reads one or more SAS files and add a \code{source_file} column
indicating the file each row came from. This column is useful for tracking
the origin of the row when combining multiple files. It also ensures that
duplicate rows across different files are not removed during the
de-duplication step in \code{convert_to_parquet()}.
}
\keyword{internal}
