% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{save_as_sas}
\alias{save_as_sas}
\title{Save a list of data frames as SAS files}
\usage{
save_as_sas(data_list, path)
}
\arguments{
\item{data_list}{A named list of data frames.}

\item{path}{Directory to save the SAS files to.}
}
\value{
\code{path}, invisibly.
}
\description{
This helper function is used for testing fastreg code and in the docs.
It will write each element of a named list as a SAS file to the given
directory. The file names are determined from the list names.
}
\examples{
save_as_sas(
  data_list = simulate_register("bef", "2020"),
  path = fs::path_temp()
)
}
