% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_register}
\alias{simulate_register}
\title{Simulate example registers}
\usage{
simulate_register(register, year = "", n = 1000)
}
\arguments{
\item{register}{Name of the register. Has to be a register accepted by
osdc::simulate_registers().}

\item{year}{Year of the register as a character. Is appended to the register
name to form the list element names.}

\item{n}{Number of rows to simulate per year.}
}
\value{
A named list of tibble(s) with names in the format
\verb{\{register\}_\{year\}}.
}
\description{
The data is simulated using \code{osdc::simulate_registers()}. It's used
in vignettes and tests.
}
\examples{
simulate_register(register = "kontakter", year = c("1999", "2000"))
}
