% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-to-sas.R
\name{simulate_registers_nested_tbl}
\alias{simulate_registers_nested_tbl}
\title{Simulate example registers along with output paths for SAS files}
\usage{
simulate_registers_nested_tbl(
  registers,
  years = "",
  n = 1000,
  output_dir = fs::path_temp("E/rawdata/701010/")
)
}
\arguments{
\item{registers}{Name of one or more registers. Must be a register that
\code{\link[osdc:simulate_registers]{osdc::simulate_registers()}} can simulate. See \code{\link[osdc:registers]{osdc::registers()}} for
a list of available registers.}

\item{years}{One or more years to save the simulated data under. The year is
used as a suffix in the file name. For example for register "bef" and year
"1999", the file will be named \code{bef1999.sas7bdat}. Can also take no year.}

\item{n}{Number of rows of data to simulate per year.}

\item{output_dir}{The root directory to append to the create SAS paths.
By default will create a path that mimics what is on DST by adding
\code{E/rawdata/701010} to the end of the output directory. The default should
technically be \verb{E:} on Windows, but the default temporary directory on
Windows for R can't allow using \code{:}, so we use \code{E} instead.}
}
\value{
A nested tibble with a column \code{data} containing the simulated data
and a column \code{output_path} containing the path where the SAS file should
be saved to. Pipe to \code{purrr::pwalk(write_to_sas)} or \code{purrr::pmap(write_to_sas)}
to write each simulated dataset to a SAS file.
}
\description{
This is a helper function that simulates data using
\code{\link[osdc:simulate_registers]{osdc::simulate_registers()}}. It's used in vignettes and tests.
It simulates data for one or more registers and years.
}
\examples{
sim_regs <- simulate_registers_nested_tbl(
  registers = c("bef", "lmdb"),
  years = c("1999", "2000"),
  n = 10,
)
sim_regs

sim_regs |>
  purrr::pwalk(write_to_sas)
}
