% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-targets.R
\name{use_targets_template}
\alias{use_targets_template}
\title{Use a targets pipeline template for converting SAS registers to Parquet}
\usage{
use_targets_template(path = "_targets.R", open = rlang::is_interactive())
}
\arguments{
\item{path}{Path to the file to create. Defaults to \verb{_targets.R}.}

\item{open}{Whether to open the file for editing. Defaults to \code{TRUE} in
interactive sessions.}
}
\value{
The path to the created file (invisibly).
}
\description{
Copies a template to your project root.
}
