% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-to-sas.R
\name{write_to_sas}
\alias{write_to_sas}
\title{Write simulated data to a SAS file}
\usage{
write_to_sas(data, output_path)
}
\arguments{
\item{data}{A tibble containing the simulated data.}

\item{output_path}{A string of the path to where the SAS file should be saved.}
}
\value{
Invisibly gives the path to the saved SAS file.
}
\description{
A helper function that writes a data frame to a SAS file. It's used
mainly in fastreg's vignettes and tests. Pipe the output of
\code{\link[=simulate_registers_with_paths]{simulate_registers_with_paths()}} with \code{\link[purrr:pwalk]{purrr::pwalk()}} followed by this function
to write each simulated dataset to a SAS file.
}
